requirements.txt
APScheduler==3.9.1.post1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gradescopecalendar==0.0.8
httplib2==0.21.0
icalendar==5.0.4
idna==3.4
oauthlib==3.2.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
